requirements.txt
python-dotenv~=1.0.1
flet~=0.23.2
beautifulsoup4~=4.12.3
requests~=2.32.3
aiogram~=3.12.0
SQLAlchemy~=2.0.32